+++ gsa/src/web/entities/page.js
@@ -16,7 +16,6 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-
 import React from 'react';
 
 import {connect} from 'react-redux';
@@ -184,6 +183,7 @@ class EntitiesPage extends React.Component {
       filterEditDialog,
       filters,
       isLoading,
+      isLoadingFilters,
       powerfilter = PowerFilter,
       onError,
       onFilterChanged,
@@ -207,6 +207,7 @@ class EntitiesPage extends React.Component {
           filter={filter}
           filters={filters}
           isLoading={isLoading}
+          isLoadingFilters={isLoadingFilters}
           onEditClick={handler}
           onError={onError}
           onRemoveClick={onFilterRemoved}
@@ -292,6 +293,7 @@ EntitiesPage.propTypes = {
   filters: PropTypes.array,
   filtersFilter: PropTypes.filter,
   isLoading: PropTypes.bool,
+  isLoadingFilters: PropTypes.bool,
   loadFilters: PropTypes.func.isRequired,
   powerfilter: PropTypes.componentOrFalse,
   section: PropTypes.componentOrFalse,
@@ -318,13 +320,16 @@ const mapStateToProps = (state, {filtersFilter}) => {
   if (!isDefined(filtersFilter)) {
     return {
       filters: [],
+      isLoadingFilters: false,
     };
   }
 
   const filterSelector = selector(state);
   const filters = filterSelector.getAllEntities(filtersFilter);
+
   return {
     filters: hasValue(filters) ? filters : [],
+    isLoadingFilters: filterSelector.isLoadingAllEntities(filtersFilter),
   };
 };